% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvictBehavior-package.R
\docType{package}
\name{ConvictBehavior-package}
\alias{ConvictBehavior}
\alias{ConvictBehavior-package}
\title{ConvictBehavior: Convict Cichlid Behavior}
\description{
The functions included in this package analyze data from experiments on social state and behavior in convict cichlids. One function performs an ANOVA and post-hoc Tukey HSD test, another checks for normality, a third graphs results as boxplots, and a fourth shows what the graph looks for different types of colorblind vision.
}
\author{
\strong{Maintainer}: Sarah Muh \email{sam9464@my.utexas.edu}

Authors:
\itemize{
  \item Maxximus Ramsaroop \email{mgramsaroop@utexas.edu}
  \item Jennie DeVore \email{jennie.devore@austin.utexas.edu}
}

}
\keyword{internal}
