% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvictFunctions.R
\name{NormCheckLog}
\alias{NormCheckLog}
\title{NormCheckLog: A function to check (log-transformed) assumptions of normality.}
\usage{
NormCheckLog(df, response, predictor1, predictor2, predictor3, type)
}
\arguments{
\item{df}{dataframe}

\item{response}{response variable (e.g., PC_1_value)}

\item{predictor1}{predictor variable (e.g., Treatment)}

\item{predictor2}{predictor variable (e.g., State)}

\item{predictor3}{predictor variable (e.g., Response)}

\item{type}{subject type (e.g., Sex)}
}
\description{
This function displays a QQ Plot of the residuals of the model and performs a Shapiro-Wilk normality test. This time, the response variable is log-transformed.
}
\examples{
NormCheckLog(df = ConvictData, response = 'PC_1_value', predictor1 = 'Treatment', predictor2 = 'State', predictor3 = 'Response', type = 'Sex')
}
\keyword{ANOVA,}
\keyword{assumptions,}
\keyword{log-transformed}
\keyword{model}
\keyword{normality,}
