% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvictFunctions.R
\name{PC2ModelSelect}
\alias{PC2ModelSelect}
\title{PC2ModelSelect: A function to select the best model.}
\usage{
PC2ModelSelect(df, response, predictor1, predictor2, predictor3, type)
}
\arguments{
\item{df}{dataframe}

\item{response}{response variable (e.g., PC_2_value)}

\item{predictor1}{predictor variable (e.g., Treatment)}

\item{predictor2}{predictor variable (e.g., State)}

\item{predictor3}{predictor variable (e.g., Response)}

\item{type}{subject type (e.g., Sex)}
}
\description{
This function assists with model selection, finding which predictor variables best explain the PC2 values.
}
\examples{
PC2ModelSelect(df = ConvictData, response = 'PC_2_value', predictor1 = 'Treatment', predictor2 = 'State', predictor3 = 'Response', type = 'Sex')
}
\keyword{ANOVA,}
\keyword{assumptions,}
\keyword{model}
\keyword{normality}
